commit includes mockups of 2 potential capstone projects I like, that still need to be proposed to the group.
Full disclosure, the mockups were generated by ChatGPT based on my prompt for what I want the final product to look like, but there is still a lot to be done-- values are hardcoded here when I want them to be entered by the user, they need to be stored in a DB, I want to change the look, add more ways to sort, keep things secure with logins/accounts, etc.

But very roughly, these are 2 ideas. If I/we come up with more ideas, they'll get their own mockups and will end up here. We'll also store the code we actually end up delivering in this repo.